+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## What's Changed
 
+- Add is_service param to inspector
 - Add inspector to load_data_point_value
 
 # Version 2025.12.12 (2025-12-07)